[benc/a11y-metrics] Rename repository stuff
